test: split text/vector RBAC e2e into two files
Single 300-line test file was hard to read. Splits into text_search_fga_tests.py
and vector_search_fga_tests.py behind two workload entries in workloads.yaml,
sharing the same setup anchor.

Also drops the multi-paragraph preamble above the Precheck helpers.

Text search RBAC: 17/17 passing. Vector search RBAC: 12/12 passing.
